# openapi/core.yaml
openapi: 3.0.3
info:
  title: CRM Core API
  description: Core API for CRM system handling Auth, Leads, Interactions, Appointments, and Payments
  version: 1.0.0
  contact:
    name: Backend Team
    email: backend@crm.com
servers:
  - url: http://localhost:3002
    description: Development server
  - url: https://api.crm.com
    description: Production server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required:
        - error
        - message
        - statusCode
      properties:
        error:
          type: string
          example: "BAD_REQUEST"
        message:
          type: string
          example: "Invalid input data"
        statusCode:
          type: integer
          example: 400
        details:
          type: object
          additionalProperties: true
...
    User:
      type: object
      required:
        - id
        - email
        - name
        - role
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        email:
          type: string
          format: email
          example: "user@example.com"
        name:
          type: string
          example: "John Doe"
        role:
          type: string
          enum: [ADMIN, MANAGER, SALES, SUPPORT]
          example: "SALES"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    Lead:
      type: object
      required:
        - id
        - name
        - email
        - phone
        - status
        - source
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174001"
        name:
          type: string
          example: "Jane Smith"
        email:
          type: string
          format: email
          example: "jane@example.com"
        phone:
          type: string
          example: "+1234567890"
        status:
          type: string
          enum: [NEW, CONTACTED, QUALIFIED, PROPOSAL, NEGOTIATION, CLOSED_WON, CLOSED_LOST]
          example: "NEW"
        source:
          type: string
          enum: [WEBSITE, SOCIAL_MEDIA, REFERRAL, COLD_CALL, EMAIL_CAMPAIGN, OTHER]
          example: "WEBSITE"
        score:
          type: integer
          minimum: 0
          maximum: 100
          nullable: true
          example: 75
        ownerId:
          type: string
          format: uuid
          nullable: true
          example: "123e4567-e89b-12d3-a456-426614174000"
        owner:
          $ref: '#/components/schemas/User'
        notes:
          type: string
          nullable: true
          example: "Interested in premium package"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
...
    Appointment:
      type: object
      required:
        - id
        - leadId
        - userId
        - title
        - scheduledAt
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174003"
        leadId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174001"
        userId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        title:
          type: string
          example: "Product Demo"
        description:
          type: string
          nullable: true
          example: "Demo of premium features"
        scheduledAt:
          type: string
          format: date-time
          example: "2024-01-20T14:00:00Z"
        status:
          type: string
          enum: [SCHEDULED, COMPLETED, CANCELLED, NO_SHOW]
          example: "SCHEDULED"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        user:
          $ref: '#/components/schemas/User'
        lead:
          $ref: '#/components/schemas/Lead'
...
paths:
...
  /api/core/leads/{id}/appointments:
    post:
      tags:
        - Appointments
      summary: Create appointment
      description: Create a new appointment for a lead
      parameters:
        - name: id
          in: path
          required: true
          description: Lead ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - scheduledAt
              properties:
                title:
                  type: string
                  example: "Product Demo"
                description:
                  type: string
                  nullable: true
                  example: "Demo of premium features"
                scheduledAt:
                  type: string
                  format: date-time
                  example: "2024-01-20T14:00:00Z"
      responses:
        '201':
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/core/appointments/{id}:
    patch:
      tags:
        - Appointments
      summary: Update appointment
      description: Update appointment status
      parameters:
        - name: id
          in: path
          required: true
          description: Appointment ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [SCHEDULED, COMPLETED, CANCELLED, NO_SHOW]
                  example: "COMPLETED"
      responses:
        '200':
          description: Appointment updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Appointment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...